ai-agent-service/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-dotenv==1.1.1
pydantic==2.11.9
requests==2.32.5
google-genai==1.41.0